pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "nix-helpers"
version = "2.1.0"
dependencies = [
    "ruamel.yaml",
    "rich",
    "podman",
    "pydantic",
]

[project.scripts]
py_motd = "py_motd.__main__:main"
sops_podman = "sops_podman.__main__:main"
py_backup = "py_backup.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
